requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
databases==0.6.1
distlib==0.3.5
Django==3.2.16
djangorestframework==3.13.1
dnspython==2.2.1
docopt==0.6.2
email-validator==1.2.1
emoji==2.2.0
env==0.1.0
environ==1.0
filelock==3.8.0
geojson==2.5.0
greenlet==1.1.3
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
mpmath==1.2.1
opencv-python==4.6.0.66
orjson==3.8.0
pbr==5.10.0
pdf2image==1.16.2
Pillow==9.2.0

